Fix out-of-tree builds
Allow out-of-tree builds using configure and/or the make.mvapich.*
scripts. This allows building for multiple target network types
from the same source tree by running the configure or make.mvapich.*
script from another directory, e.g.:

 mkdir -p builds/gen2
 cd builds/gen2 && ../../make.mvapich.gen2

This patch fixes up the make.mvapich* scripts by referencing
configure by a relative path (instead of ./) and fixes up the
channel-specific Makefile.in files by ensuring that all source files
are referenced relative to $(srcdir), and build products are
referenced relative to the build dir.

The 'process' subdirectory of each channel is handled specially since
it is not generated by ./configure (thus the directory isn't even
created in the builddir). It may be better to generate the Makefile
with configure in the future, but for now this nominally works.